# api/swagger.yaml
swagger: "2.0"
info:
  description: "Play, browse and explore World's music in one place."
  version: "Pre-alpha"
  title: "Canzona"
  contact:
    email: "anur@blink.ba"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "canzona.io"
basePath: "/"
tags:
- name: "User"
  description: "User routes"
schemes:
- "http" # local env
- "https"
paths:
  /users/register:
    post:
      tags:
      - "User"
      summary: "Register a new user"
      operationId: "register"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body of the request"
        required: true
        schema:
          $ref: "#/definitions/UserRegister"
      responses:
        200:
          description: "Success"
        429:
          description: "Request rate limit exceeded"
        400:
          description: "Bad request"
  /users/login:
    post:
      tags:
      - "User"
      summary: "Login an existing user"
      operationId: "login"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body of the request"
        required: true
        schema:
          $ref: "#/definitions/UserLogin"
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /users/authenticate:
    post:
      tags:
      - "User"
      summary: "Authenticate the user"
      operationId: "authenticate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - name: "Authorization"
          type: "string"
          in: "header"
          description: "Unique JSONWebToken for authentication of the user"
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /users/recover:
    post:
      tags:
      - "User"
      summary: "Recover the user account"
      operationId: "recover"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Body of the request"
          required: true
          schema:
            properties:
              email:
                type: "string"
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
        404:
          description: "Account doesn't exist"
        429:
          description: "Request rate limit exceeded"
        500:
          description: "Internal server error"
  /users/recover/conclude:
    put:
      tags:
      - "User"
      summary: "Update the password of the account and regain the access of it"
      operationId: "concludeAccountRecovery"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization"
          description: "Unique JSONWebToken for authentication of the user"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          description: "Body of the request"
          required: true
          schema:
            properties:
              password:
                type: "string"
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
        404:
          description: "Account doesn't exist"
        429:
          description: "Request rate limit exceeded"
        500:
          description: "Internal server error"
  /artists/create:
    post:
      tags:
      - "Artist"
      summary: "Create a new artist"
      operationId: "createArtist"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - name: "Authorization"
          type: "string"
          in: "header"
          description: "Unique JSONWebToken for authentication of the admin"
        - in: "body"
          name: "body"
          description: "Body of the request"
          required: true
          schema:
            $ref: "#/definitions/Artist"
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "string"
      username:
        type: "string"
      name:
        type: "string"
      surname:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
        format: "password"
      verified:
        type: "boolean"
      avatar:
        type: "string"
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
  UserRegister:
    type: "object"
    properties:
      username:
        type: "string"
      name:
        type: "string"
      surname:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
        format: "password"
  UserLogin:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
        format: "password"

  Artist:
    type: "object"
    properties:
      name:
        type: "string"
      isGroup:
        type: "boolean"
      formed:
        type: "string"
        format: "date-time"
      avatar:
        type: "string"
      background:
        type: "string"
      monthlyListeners:
        type: "integer"
      about:
        type: "string"
      gallery:
        type: "array"
        items:
          type: "string"

externalDocs:
  description: "Official docs and repository available"
  url: "https://github.com/bashovski/canzona"